Break up the main test a bit
